requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
torch>=2.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
tiktoken>=0.5.0
rich>=13.0.0

# Optional dependencies for small models and quantization
bitsandbytes>=0.41.0  # For 8-bit quantization
optimum>=1.14.0       # For model optimization
